# src/track/dto/create-track.dto.ts
import { IsOptional, IsString, IsUUID, IsInt } from 'class-validator';

export class CreateTrackDto {
  @IsString()
  name: string;

  @IsOptional()
  @IsUUID('4')
  artistId: string | null;

  @IsOptional()
  @IsUUID('4')
  albumId: string | null;

  @IsInt()
  duration: number;
}